Added tests and refactored
ModuleComputerCore is now less of a god class because it now uses the
StackMachine and Bytecode classes to manage the computer parts. All
editor window logic is now delegated to the ScriptingInterface class.
There's some problem where ksp isn't loading parts when the mod is
installed, and I think it might have to do with the load/save methods or
the computerCore.cfg, so I'm looking through those for some sort of
cause, but don't have a solution yet.